requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.0
tzdata
convertdate
psycopg2-binary==2.9.9
pandas>=2.0.0
openpyxl>=3.1.0

# Testing dependencies (optional)
pytest==8.3.2
pytest-cov==5.0.0

# Performance monitoring (optional)
psutil==5.9.8

